+++ logs/logs.go
@@ -49,6 +49,22 @@ type Log struct {
 	hasLogged bool
 }
 
+// TraceID returns the Trace ID of the log
+func (l *Log) TraceID() string {
+	if l == nil {
+		return ""
+	}
+	return l.traceID
+}
+
+// SpanID returns the Span ID of the log
+func (l *Log) SpanID() string {
+	if l == nil {
+		return ""
+	}
+	return l.spanID
+}
+
 // NewLog is a constructor for a log object
 func (l *Logger) NewLog(traceID string, request RequestContext) *Log {
 	if traceID == "" {
@@ -108,7 +124,18 @@ func (l *Logger) NewRequestLog(r *http.Request) *Log {
 
 // SetContext sets the provided context key to the provided value
 func (l *Log) SetContext(fieldName string, value interface{}) {
+	if l == nil {
+		return
+	}
 	l.context[fieldName] = value
+}
+
+// GetContext gets the provided context key
+func (l *Log) GetContext(fieldName string) interface{} {
+	if l == nil {
+		return nil
+	}
+	return l.context[fieldName]
 }
 
 // AddContext adds any relevant unstructured data to context map