+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Expose access to log trace ID, span ID, and context [#28](https://github.com/rokwire/logging-library-go/issues/28)
+
 ## [2.2.0] - 2023-01-19
 ### Added
 - Handle HTTP Cookies [#25](https://github.com/rokwire/logging-library-go/issues/25)